Moved logic to SynteticImageLoader
